[Rest-server] Rest-server update for RBAC mode (#3433)
* Update user migrate script. And separate into 2.

* Update user migrate script. And separate into 2.

* Update user migrate script. And separate into 2.

* [Rest-server] secret filter (#3441)

* Remove unused value.

* [Rest-server] CA and TOKEN issue fix. (#3452)

* [rest-server] Permission Update (#3462)
