+++ src/rest-server/deploy/rbac.yaml
@@ -79,3 +79,20 @@ subjects:
 - kind: ServiceAccount
   name: rest-server-openpai
   namespace: default
+
+---
+
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRoleBinding
+metadata:
+  name: rest-server-openpai-edit
+roleRef:
+  apiGroup: rbac.authorization.k8s.io
+  kind: ClusterRole
+  name: edit
+subjects:
+  - kind: ServiceAccount
+    name: rest-server-openpai
+    namespace: default
+
+
