+++ src/rest-server/src/models/v2/user.js
@@ -25,13 +25,13 @@ const k8sConfig = require('@pai/config/kubernetes');
 const crudType = 'k8sSecret';
 const crudUser = crudUtil.getStorageObject(crudType);
 let optionConfig = {};
-if (k8sConfig.ca) {
-  optionConfig.k8sAPIServerCaFile = k8sConfig.ca;
+if (k8sConfig.apiserver.ca) {
+  optionConfig.k8sAPIServerCaFile = k8sConfig.apiserver.ca;
 }
-if (k8sConfig.token) {
-  optionConfig.k8sAPIServerTokenFile = k8sConfig.token;
+if (k8sConfig.apiserver.token) {
+  optionConfig.k8sAPIServerTokenFile = k8sConfig.apiserver.token;
 }
-const crudConfig = crudUser.initConfig(process.env.K8S_APISERVER_URI, optionConfig);
+const crudConfig = crudUser.initConfig(k8sConfig.apiserver.uri, optionConfig);
 
 // crud user wrappers
 const getUser = async (username) => {
