+++ src/rest-server/src/utils/manager/group/crudK8sSecret.js
@@ -17,8 +17,8 @@
 
 const Group = require('./group');
 const axios = require('axios');
-const {readFileSync} = require('fs');
 const {Agent} = require('https');
+const logger = require('@pai/config/logger');
 
 /**
  * @typedef Config
@@ -57,11 +57,11 @@ function initConfig(apiServerUri, option = {}) {
     },
   };
   if ('k8sAPIServerCaFile' in option) {
-    const ca = readFileSync(option.k8sAPIServerCaFile);
+    const ca = option.k8sAPIServerCaFile;
     config.requestConfig.httpsAgent = new Agent({ca});
   }
   if ('k8sAPIServerTokenFile' in option) {
-    const token = readFileSync(option.k8sAPIServerTokenFile, 'ascii');
+    const token = option.k8sAPIServerTokenFile;
     config.requestConfig.headers = {Authorization: `Bearer ${token}`};
   }
   return config;
@@ -117,13 +117,17 @@ async function readAll(config) {
     let allGroupInstance = [];
     let groupData = response['data'];
     for (const item of groupData['items']) {
-      let groupInstance = Group.createGroup({
-        'groupname': Buffer.from(item['data']['groupname'], 'base64').toString(),
-        'description': Buffer.from(item['data']['description'], 'base64').toString(),
-        'externalName': Buffer.from(item['data']['externalName'], 'base64').toString(),
-        'extension': JSON.parse(Buffer.from(item['data']['extension'], 'base64').toString()),
-      });
-      allGroupInstance.push(groupInstance);
+      try {
+        let groupInstance = Group.createGroup({
+          'groupname': Buffer.from(item['data']['groupname'], 'base64').toString(),
+          'description': Buffer.from(item['data']['description'], 'base64').toString(),
+          'externalName': Buffer.from(item['data']['externalName'], 'base64').toString(),
+          'extension': JSON.parse(Buffer.from(item['data']['extension'], 'base64').toString()),
+        });
+        allGroupInstance.push(groupInstance);
+      } catch (error) {
+        logger.debug(`secret ${item['metadata']['name']} is filtered in ${config.namespace} due to group schema`);
+      }
     }
     return allGroupInstance;
   } catch (error) {
