+++ src/rest-server/src/utils/manager/user/crudK8sSecret.js
@@ -17,8 +17,8 @@
 
 const User = require('./user');
 const axios = require('axios');
-const {readFileSync} = require('fs');
 const {Agent} = require('https');
+const logger = require('@pai/config/logger');
 
 /**
  * @typedef Config
@@ -57,11 +57,11 @@ function initConfig(apiServerUri, option = {}) {
     },
   };
   if ('k8sAPIServerCaFile' in option) {
-    const ca = readFileSync(option.k8sAPIServerCaFile);
+    const ca = option.k8sAPIServerCaFile;
     config.requestConfig.httpsAgent = new Agent({ca});
   }
   if ('k8sAPIServerTokenFile' in option) {
-    const token = readFileSync(option.k8sAPIServerTokenFile, 'ascii');
+    const token = option.k8sAPIServerTokenFile;
     config.requestConfig.headers = {Authorization: `Bearer ${token}`};
   }
   return config;
@@ -118,14 +118,18 @@ async function readAll(config) {
     let allUserInstance = [];
     let userData = response['data'];
     for (const item of userData['items']) {
-      let userInstance = User.createUser({
-        'username': Buffer.from(item['data']['username'], 'base64').toString(),
-        'password': Buffer.from(item['data']['password'], 'base64').toString(),
-        'grouplist': JSON.parse(Buffer.from(item['data']['grouplist'], 'base64').toString()),
-        'email': Buffer.from(item['data']['email'], 'base64').toString(),
-        'extension': JSON.parse(Buffer.from(item['data']['extension'], 'base64').toString()),
-      });
-      allUserInstance.push(userInstance);
+      try {
+        let userInstance = User.createUser({
+          'username': Buffer.from(item['data']['username'], 'base64').toString(),
+          'password': Buffer.from(item['data']['password'], 'base64').toString(),
+          'grouplist': JSON.parse(Buffer.from(item['data']['grouplist'], 'base64').toString()),
+          'email': Buffer.from(item['data']['email'], 'base64').toString(),
+          'extension': JSON.parse(Buffer.from(item['data']['extension'], 'base64').toString()),
+        });
+        allUserInstance.push(userInstance);
+      } catch (error) {
+        logger.debug(`secret ${item['metadata']['name']} is filtered in ${config.namespace} due to user schema`);
+      }
     }
     return allUserInstance;
   } catch (error) {
